Fix(Graphics): switch to 16-bit RGB565 and QNX-verified HDMI config
Root cause: BBB SRM 6.6.3 confirms only 16 lcd_data pins wired to
TDA19988 (VP[23:16] tied low). Previous 24bpp unpacked mode sent data
on all 24 pins but TDA only received lower 16 — complete data mismatch.

LCDC changes:
- Switch from 24bpp XRGB8888 to 16-bit RGB565 (2 bytes/pixel)
- Remove TFT_24BPP_MODE and TFT_24BPP_UNPACK flags from RASTER_CTRL
- Remove GPMC_AD[8:15] pinmux (pins go to expansion header, not TDA)
- Add FIFO threshold 512 (QNX production value) to prevent underflow
- Add CLKC_RESET sequence before configuration
- Add SYNC_EDGE + HSYNC_INVERT in TIMING_2 (QNX polarity config)
- Remove FDD=128 (QNX uses FDD=0 to prevent DMA starvation)
- Fix HORMSB >>3 (hardware verified: bit 2 reserved on AM335x)
- Framebuffer API returns uint16_t* instead of uint32_t*

TDA19988 changes:
- VIP muxing switched to QNX 16-bit values (CNTRL_1=0x01, CNTRL_2=0x45)
- TBG config: remove X_EXT/H_EXT/V_EXT (QNX uses only TGL_EN)
- Timing values recalculated using QNX formula (REFPIX=153, REFLINE=6,
  DE=370-1650, VWIN=29-749 vs old NXP BSL values)
- Init order: dpms (ports+mux) before mode_set (QNX sequence)
- REG_ENABLE_SPACE = 0x01 (QNX value, was 0x00)
- TBG_CNTRL_0: clear SYNC_MTHD then SYNC_ONCE as last register
- Add CCLK_ON and explicit FEAT_POWERDOWN management
- Remove PLL writes from probe (deferred to post_lcdc_init)
- Restore MAIN_CNTRL0_SR reset (all production drivers use it)

+++ VinixOS/kernel/include/lcdc.h
@@ -40,9 +40,9 @@ void lcdc_init(void);
 
 /**
  * Get pointer to framebuffer pixel data (after palette header)
- * @return Kernel VA pointer to first pixel
+ * @return Kernel VA pointer to first pixel (RGB565, uint16_t per pixel)
  */
-uint32_t *lcdc_get_framebuffer(void);
+uint16_t *lcdc_get_framebuffer(void);
 
 /**
  * Get display width in pixels
